SHIP-3308 Library function to find the appropriate binary (#1)
Move the binary file selection logic from `get_wkhtmltopdf_bin()` of
the main ShipEarly app project to this project, so that any future
changes to the set of binaries, such as a version update, and the
matching change to the selection logic is encapsulated in this project.

- Copy `get_wkhtmltopdf_bin()` to a PHP source file in this project.
- Include the source file in the composer autoload listing so that it
can be imported to replace `get_wkhtmltopdf_bin()` in the main project.
- Namespace the function so that it is not imported as a pure global.